term값을 String으로 intent 받으면 rate로 넘어감
