+++ app/src/main/java/kr/co/gachon/emotion_diary/ui/Remind/remindSummary/RemindSummaryActivity.java
@@ -1,7 +1,11 @@
 package kr.co.gachon.emotion_diary.ui.Remind.remindSummary;
 
 import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageButton;
+import android.widget.TextView;
 
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.Date;
@@ -11,7 +15,7 @@ import kr.co.gachon.emotion_diary.ui.Remind.WriteRate.RateFragment;
 import kr.co.gachon.emotion_diary.ui.Remind.emotionStatistics.EmotionStatisticsFragment;
 import kr.co.gachon.emotion_diary.ui.Remind.timeGraph.TimeZoneFragment;
 
-public class RemindSummary extends AppCompatActivity {
+public class RemindSummaryActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -45,5 +49,25 @@ public class RemindSummary extends AppCompatActivity {
                 .replace(R.id.summaryFragmentEmotion,  emotionFragment)
                 .replace(R.id.summaryFragmentTime, timeFragment)
                 .commit();
+
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
+            actionBar.setCustomView(R.layout.custom_back_bar);
+
+            ImageButton backButton = actionBar.getCustomView().findViewById(R.id.backButtonActionBar);
+            backButton.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    finish();
+                }
+            });
+
+            // 액션 바 제목 바꾸기
+            TextView titleTextView = actionBar.getCustomView().findViewById(R.id.titleTextViewActionBar);
+            if (titleTextView != null) {
+                titleTextView.setText("Remind Summary");
+            }
+        }
     }
 }
