+++ app/src/main/java/kr/co/gachon/emotion_diary/ui/Remind/timeGraph/TimeZoneActivity.java
@@ -2,20 +2,20 @@ package kr.co.gachon.emotion_diary.ui.Remind.timeGraph;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.FragmentTransaction;
 
 import java.util.Date;
 
 import kr.co.gachon.emotion_diary.R;
-import kr.co.gachon.emotion_diary.ui.Remind.WriteRate.RateActivity;
-import kr.co.gachon.emotion_diary.ui.Remind.emotionStatistics.EmotionStatisticsActivity;
-import kr.co.gachon.emotion_diary.ui.Remind.emotionStatistics.EmotionStatisticsFragment;
-import kr.co.gachon.emotion_diary.ui.Remind.remindSummary.RemindSummary;
+import kr.co.gachon.emotion_diary.ui.Remind.remindSummary.RemindSummaryActivity;
 
 public class TimeZoneActivity extends AppCompatActivity {
     private TextView timeTextView;
@@ -45,12 +45,32 @@ public class TimeZoneActivity extends AppCompatActivity {
 
         Button button = findViewById(R.id.nextButton);
         button.setOnClickListener(view -> {
-            Intent intent = new Intent(TimeZoneActivity.this, RemindSummary.class);
+            Intent intent = new Intent(TimeZoneActivity.this, RemindSummaryActivity.class);
             intent.putExtra("isMonthly", isMonthly);
             intent.putExtra("startDate", startDate);
             intent.putExtra("endDate", endDate);
             startActivity(intent);
         });
+
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
+            actionBar.setCustomView(R.layout.custom_back_bar);
+
+            ImageButton backButton = actionBar.getCustomView().findViewById(R.id.backButtonActionBar);
+            backButton.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    finish();
+                }
+            });
+
+            // 액션 바 제목 바꾸기
+            TextView titleTextView = actionBar.getCustomView().findViewById(R.id.titleTextViewActionBar);
+            if (titleTextView != null) {
+                titleTextView.setText("Time Zone");
+            }
+        }
     }
 
     public void setTimeText(String time) {
