+++ app/src/main/res/layout/custom_back_bar.xml
@@ -1,24 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal">
+    android:layout_height="?attr/actionBarSize"
+    android:gravity="center_vertical">
 
     <ImageButton
         android:id="@+id/backButtonActionBar"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_height="match_parent"
         android:background="?attr/selectableItemBackgroundBorderless"
-        android:layout_gravity="start"
+        android:layout_alignParentStart="true"
+        android:layout_centerVertical="true"
         android:src="@drawable/left"/>
 
     <TextView
         android:id="@+id/titleTextViewActionBar"
-        android:layout_width="100dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginStart="100dp"
-        android:text="@string/diary_write_title"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:layout_toStartOf ="@id/backButtonActionBar"
+        android:layout_alignParentEnd="true"
+        android:layout_centerVertical="true"
         android:textSize="18dp" />
 
-</LinearLayout>
+</RelativeLayout>
